+++ lib/app.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+
+class MyApp extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'BudgetBuddy',
+      theme: ThemeData(
+        primarySwatch: Colors.purple,
+        visualDensity: VisualDensity.adaptivePlatformDensity,
+      ),
+      home: Scaffold(
+        appBar: AppBar(
+          title: const Text('BudgetBuddy'),
+        ),
+        body: const Center(
+          child: Text('Hello, World!'),
+        ),
+      ),
+    );
+  }
+}
