+++ lib/main.dart
@@ -1,26 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:budget_buddy/app.dart';
 
 void main() {
   runApp(MyApp());
 }
 
-class MyApp extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'BudgetBuddy',
-      theme: ThemeData(
-        primarySwatch: Colors.purple,
-        visualDensity: VisualDensity.adaptivePlatformDensity,
-      ),
-      home: Scaffold(
-        appBar: AppBar(
-          title: Text('BudgetBuddy'),
-        ),
-        body: Center(
-          child: Text('Hello, World!'),
-        ),
-      ),
-    );
-  }
-}
