+++ lib/router/router.dart
@@ -1,0 +1,5 @@
+import '../screens/start_screen.dart';
+
+final routes = {
+  '/': (context) => const StartScreen(title: 'Start Screen'),
+};
