+++ lib/screens/start_screen.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class StartScreen extends StatefulWidget {
+  const StartScreen({super.key, required String title});
+
+  @override
+  State<StartScreen> createState() => _StartScreenState();
+}
+
+class _StartScreenState extends State<StartScreen> {
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
